python_app/requirements.txt
customtkinter
pyserial
SpeechRecognition
pyaudio
pyttsx3
opencv-python
gTTS
pygame
